Reporting Service/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
PyJWT==2.8.0
httpx==0.27.2
reportlab==4.1.0
